backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.0.3
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0.post1 
openai==1.3.7
pydantic-settings==2.1.0 